Optimized project structure and updated associated paths
Adjustments were made to the project's structure and corresponding file locations to enhance organization and ease of maintenance. As a result, the root in vite.config.js was altered from 'src'. Paths in the index.html and main.css files were adapted to reflect this modification. Upon changing the directory, the font paths were adjusted to link correctly. The package version was updated in both the package-lock.json and package.json files. Also, changes were made to the TailwindCSS configuration file and its paths. Lastly, the src/fonts/Outfit/OFL.txt was renamed to assets/fonts/Outfit/OFL.txt.