Refactor README and update package versions
This commit includes two main changes. Firstly, the README file was thoroughly refactored and reduced in size from 103 to 60 lines, offering more concise and relevant information about the project it describes. This change was implemented in order to make the README more reader-friendly and efficient in conveying essential information. Secondly, the different versions of the package 'rollup' listed in package-lock.json were updated from version 4.5.2 to version 4.6.0. This was done to ensure the project uses up-to-date dependencies, potentially benefiting from bug fixes, new features, and performance improvements.